refactor: f32 -> i32

+++ src/webgpu/shaderity_shaders/FlatSingleShader/FlatSingleShader.vert
@@ -29,8 +29,8 @@ fn main(
 
   var output : VertexOutput;
 
-  let cameraSID = get_currentComponentSIDs(0.0, /* shaderity: @{WellKnownComponentTIDs.CameraComponentTID} */);
-  let worldMatrix = get_worldMatrix(instance_ids.x);
+  let cameraSID = i32(get_currentComponentSIDs(0, /* shaderity: @{WellKnownComponentTIDs.CameraComponentTID} */));
+  let worldMatrix = get_worldMatrix(i32(instance_ids.x));
   let viewMatrix = get_viewMatrix(cameraSID, 0);
   let projectionMatrix = get_projectionMatrix(cameraSID, 0);
